refactor: extract selectors to separate file for recipes
